amount changed from int to double

+++ lib/products_sales/data/models/product_sales_model.dart
@@ -33,7 +33,7 @@ class ProductSalesModel extends ProductSale {
       {String? id,
       String? buyerName,
       String? productName,
-      int? amount,
+      double? amount,
       double? totalCost,
       double? paidAmount,
       double? unPaidAmount,
